refactor: Move Header component to new path and update imports
