+++ app/dashboard/loading.tsx
@@ -1,5 +1,5 @@
 import { Loading } from "@/components/loading";
-import { Header } from "@/components/editor/header";
+import { Header } from "@/components/header";
 
 export default function DashboardLoading() {
   return (
