+++ app/editor/theme/[[...themeId]]/page.tsx
@@ -2,7 +2,7 @@ import { getEditorConfig } from "@/config/editors";
 import { cn } from "@/lib/utils";
 import Editor from "@/components/editor/editor";
 import { Metadata } from "next";
-import { Header } from "../../../../components/editor/header";
+import { Header } from "@/components/header";
 import { getTheme } from "@/actions/themes";
 import { Suspense } from "react";
 import { Loading } from "@/components/loading";
