+++ app/pricing/layout.tsx
@@ -1,4 +1,4 @@
-import { Header } from "@/components/editor/header";
+import { Header } from "@/components/header";
 
 export default function PricingLayout({ children }: { children: React.ReactNode }) {
   return (
