+++ app/themes/[themeId]/loading.tsx
@@ -1,4 +1,4 @@
-import { Header } from "@/components/editor/header";
+import { Header } from "@/components/header";
 import { Loading } from "@/components/loading";
 
 export default function ThemeLoading() {
